+++ riddle-audio/src/audio_system.rs
@@ -8,7 +8,7 @@ use std::{collections::HashMap, sync::Mutex, time::Instant};
 /// way to use this type is to let the `riddle` crate initialize and manage it for you.
 ///
 /// It is possible to manage the audio system state independantly - the most important
-/// thing to note is that [`AudioSystem::process_frame`] must be called periodically
+/// thing to note is that [`ext::AudioSystemExt::process_frame`] must be called periodically
 /// for [`ClipPlayer`] to work properly. This is **not** something that needs doing if
 /// using the `riddle` crate to manage the [`AudioSystem`] automatically.
 pub struct AudioSystem {
@@ -20,8 +20,23 @@ pub struct AudioSystem {
 define_handles!(<AudioSystem>::weak_self, pub AudioSystemHandle, pub AudioSystemWeak);
 
 impl AudioSystem {
-    /// Create the audio system, connected to the default audio output device.
-    pub fn new() -> Result<(AudioSystemHandle, AudioMainThreadState)> {
+    pub(crate) fn register_fade(&self, fade: Fade) {
+        let mut fades = self.fades.lock().unwrap();
+        let existing = fades.remove(&fade.key());
+        match existing {
+            Some(old) => fades.insert(fade.key(), Fade::merge_pair(old, fade)),
+            None => fades.insert(fade.key(), fade),
+        };
+    }
+
+    fn tick_fades(&self, now: Instant) {
+        let mut fades = self.fades.lock().unwrap();
+        fades.retain(|_, f| f.update(now));
+    }
+}
+
+impl ext::AudioSystemExt for AudioSystem {
+    fn new() -> Result<(AudioSystemHandle, AudioMainThreadState)> {
         let (stream, stream_handle) =
             rodio::OutputStream::try_default().map_err(|_| AudioError::InitFailed {
                 cause: "Failed to get rodio output device",
@@ -37,40 +52,9 @@ impl AudioSystem {
         ))
     }
 
-    /// Update the system's state.
-    ///
-    /// Updates all [`ClipPlayer`] fades. This must be called periodically for the [`AudioSystem`]
-    /// to function. **Do not** call this if the `riddle` crate is being used.
-    ///
-    /// # Example
-    /// ```no_run
-    /// # use riddle_audio::*; fn main() -> Result<(), AudioError> {
-    /// let (audio_system, _audio_main_thread_state) = AudioSystem::new()?;
-    ///
-    /// // Tick the audio system every 100ms
-    /// let start_time = std::time::Instant::now();
-    /// while std::time::Instant::now() - start_time < std::time::Duration::from_secs(2) {
-    ///     audio_system.process_frame();
-    ///     std::thread::sleep(std::time::Duration::from_millis(100));
-    /// }
-    /// # Ok(()) }
-    /// ```
-    pub fn process_frame(&self) {
+    fn process_frame(&self) {
         let now = Instant::now();
         self.tick_fades(now);
-    }
-    pub(crate) fn register_fade(&self, fade: Fade) {
-        let mut fades = self.fades.lock().unwrap();
-        let existing = fades.remove(&fade.key());
-        match existing {
-            Some(old) => fades.insert(fade.key(), Fade::merge_pair(old, fade)),
-            None => fades.insert(fade.key(), fade),
-        };
-    }
-
-    fn tick_fades(&self, now: Instant) {
-        let mut fades = self.fades.lock().unwrap();
-        fades.retain(|_, f| f.update(now));
     }
 }