+++ riddle-audio/src/ext/mod.rs
@@ -1,0 +1,5 @@
+//! Types which are needed to use the audio system independent from the root `riddle` crate.
+
+mod audio_system;
+
+pub use audio_system::*;