+++ riddle-font/src/rusttype_ext/ttfont.rs
@@ -1,0 +1,3 @@
+pub trait RustTypeTTFontExt {
+    fn rustype_font(&self) -> &rusttype::Font<'static>;
+}