+++ riddle-image/src/image_ext/image.rs
@@ -1,0 +1,3 @@
+pub trait ImageImageExt {
+    fn image_rgbaimage(&self) -> &::image::RgbaImage;
+}