+++ riddle-image/src/image_ext/mod.rs
@@ -1,0 +1,5 @@
+//! Traits which expose underlying image crate's types
+
+mod image;
+
+pub use self::image::*;