+++ riddle-input/src/ext/mod.rs
@@ -1,0 +1,5 @@
+//! Types which are needed to use the input system independent from the root `riddle` crate.
+
+mod input_system;
+
+pub use input_system::*;