+++ riddle-platform-winit/src/ext/mod.rs
@@ -1,0 +1,5 @@
+//! Types which are needed to use the input system independent from the root `riddle` crate.
+
+mod platform_system;
+
+pub use platform_system::*;