+++ riddle-platform-winit/src/platform_system.rs
@@ -19,48 +19,6 @@ pub struct PlatformSystem {
 define_handles!(<PlatformSystem>::weak_self, pub PlatformSystemHandle, pub PlatformSystemWeak);
 
 impl PlatformSystem {
-    /// Create a new platform system and its main thread state pair.
-    ///
-    /// **Do not** call if using the `riddle` crate as recommended, as ` RiddleLib` manages
-    /// the creation and platform lifetime automatically.
-    ///
-    /// # Example
-    ///
-    /// This example is only relevant if you're not using `riddle`
-    ///
-    /// ```no_run
-    /// # use riddle_platform_winit::*;
-    /// # fn main() -> Result<(), PlatformError> {
-    /// let (platform_system, main_thread_state) = PlatformSystem::new();
-    /// let window = WindowBuilder::new().build(main_thread_state.borrow_context())?;
-    ///
-    /// main_thread_state.run::<PlatformError, _>(move |ctx| {
-    ///     match ctx.event() {
-    ///         PlatformEvent::WindowClose(_) => { ctx.quit(); }
-    ///         _ => ()
-    ///     };
-    ///     Ok(())
-    /// })
-    /// # }
-    /// ```
-    pub fn new() -> (PlatformSystemHandle, PlatformMainThreadState) {
-        let event_loop = winit::event_loop::EventLoop::with_user_event();
-        let event_proxy = event_loop.create_proxy();
-        let system = PlatformSystemHandle::new(|weak_self| PlatformSystem {
-            weak_self,
-            event_proxy: Mutex::new(event_proxy),
-
-            window_map: WindowMap::new().into(),
-
-            event_pub: EventPub::new(),
-        });
-        let main_thread_state = PlatformMainThreadState {
-            system: system.clone(),
-            event_loop: RefCell::new(Some(event_loop)),
-        };
-        (system, main_thread_state)
-    }
-
     /// Get the [`PlatformEvent`] publisher, so that other systems can consume events.
     ///
     /// # Example
@@ -112,6 +70,26 @@ impl PlatformSystem {
         for window in windows {
             window.update()
         }
+    }
+}
+
+impl ext::PlatformSystemExt for PlatformSystem {
+    fn new() -> (PlatformSystemHandle, PlatformMainThreadState) {
+        let event_loop = winit::event_loop::EventLoop::with_user_event();
+        let event_proxy = event_loop.create_proxy();
+        let system = PlatformSystemHandle::new(|weak_self| PlatformSystem {
+            weak_self,
+            event_proxy: Mutex::new(event_proxy),
+
+            window_map: WindowMap::new().into(),
+
+            event_pub: EventPub::new(),
+        });
+        let main_thread_state = PlatformMainThreadState {
+            system: system.clone(),
+            event_loop: RefCell::new(Some(event_loop)),
+        };
+        (system, main_thread_state)
     }
 }