+++ riddle-platform-winit/src/winit_ext/mod.rs
@@ -1,0 +1,5 @@
+//! Traits which expose underlying winit crate's types
+
+mod window;
+
+pub use window::*;