+++ riddle-platform-winit/src/winit_ext/window.rs
@@ -1,0 +1,3 @@
+pub trait WinitWindowExt {
+    fn winit_window(&self) -> &winit::window::Window;
+}