+++ CHANGELOG.md
@@ -15,5 +15,6 @@
 * docs: Add an example combining the riddle renderer with a custom wgpu renderer.
 * input: Add `InputSystem::gamepad_axis_value` and `InputSystem::last_active_gamepad`.
 * docs: Update pong example to support gamepad input
+* **breaking**: Put internal system methods in to Ext traits, expose underlying types with crate specific Ext traits.
 
 ## 0.1.0 - Initial Release